browser: core/Util.js -> app/Util.ts without namespaces
